Refactored components (WIP) #156
