+++ client/src/components/pam/DashboardEmbed.vue
@@ -53,7 +53,7 @@
 import _ from 'lodash'
 
 import Panel1 from './Panel1.vue'
-import ProjectViewer from '@/components/project/ProjectViewer.vue'
+import ProjectViewer from '@/pages/Project.vue'
 import CreateAsset from '@/components/pam/CreateAsset.vue'
 import FileManager from '@/components/fileManager/FileManager.vue'
