+++ client/src/layouts/Base.vue
@@ -32,7 +32,7 @@
 <script>
 import { sync } from 'vuex-pathify'
 import { openURL } from 'quasar'
-import NavBar from '../components/layout/NavBar'
+import NavBar from '@/components/layout/NavBar'
 
 export default {
   name: 'MyLayout',
